feat: 레벨 동기화 시 플레이어의 복사본과 원본을 분리해 Thread safe 하도록 함

+++ CrazyArcade/ConsoleEngine/Network/Packets.h
@@ -135,7 +135,7 @@ public:
     char* gameStateBuffer;
 
     GameStateSynchronizePacket(size_t gameStateSize, char* gameStateData)
-        : gameStateBuffer(gameStateData), gameStateSize(gameStateSize)
+        : gameStateBuffer(gameStateData), gameStateSize((uint32_t)gameStateSize)
     {
         header.packetType = (uint32_t)PacketType::GAME_STATE_SYNCHRONIZE;
         header.packetSize = sizeof(GameStateSynchronizePacket);
